Add a definition in page_layouts to render pages with certain layouts.
For pages, we have a new PERMITTED_ATTRIBUTE 'render_layout', which
allows us to define a layout to render the page with. Our use case for
this is to render certain pages with a layout that's not the
'application' layout.

In the pages controller, we check to see if the page has a defined
render_layout. If so, we get the render_layout, and call render with the
render_layout. If not, we just do whatever we did before.

natures.rb contains utility methods for page_layouts, so I've added
utility functions 'has_render_layout' and 'render_layout' there.