Modified the CMS to support AMP pages
We have generally have two separate files for each component of the CMS (from the page layout down to the essence), depending on whether the AMP version is different. The two files are differentiated by their format or extension, and we decide which of the two serve depending on the format appended to the end of page path. All AMP pages will have a page path ending in .amp. For instance, /wedding-invitations.amp or /holiday-christmas-cards.html.amp and the corresponding page layout name will be application.amp.erb (rather than application.html.erb). From there on, we explicitly pass the format down to the components on the page so that the correct version is rendered.

- Check the format (page path extension) to decide which version (AMP or HTML) of the main layout page to serve
- Used the previously supported 'render_format' variable passed in through options to decide which version (AMP or HTML) of the cell, element, etc. to render

https://app.asana.com/0/380306915042955/list