CategorySettings.java will parse contents for data.
Alchemy CMS separated contents from the page data, so that obtaining the content data requires at least 2 API calls. To keep only 1 API call, we add contents back into the page JSON data.

NOTE: In the future, we would like to switch to the LegacyElementSerializer based on a parameter we pass into the URL.