Allow a custom select2 template for essence select
This is done by having a string of the template function that the
template needs to call as a data attribute on the essence select editor.
To support this we now pass the data attributes when creating the select
tags. When Alchemy initializes the select2 elements, we look for select2
elements with that particular template data attribute, and instead of
initializing it as usual, we use a template specified in that the
template data attribute instead.

Additinally, Alchemy previously would generate select values as an
optgroup if we passed in a hash. But to allow for a more flexible label
as the select option, we added in a varaible 'use_opt_group' which if
true will generate an optgroup, rather than options.

https://app.asana.com/0/507541520661760/709944895722331